+++ playlet-web/src/lib/Types.ts
@@ -1,5 +1,10 @@
 export const ScreenNames = ["search", "home", "bookmarks", "settings", "info", "remote"] as const;
 
+export enum ProfileAuthState {
+    Authenticated = "Authenticated",
+    NeedsReauth = "NeedsReauth",
+}
+
 export type AppState = {
     screen: "search" | "home" | "bookmarks" | "settings" | "info" | "remote"
 }